+++ src/mod_perimeterx.c
@@ -20,7 +20,6 @@
 #include <http_request.h>
 #include <http_log.h>
 #include <apr_strings.h>
-#include <apr_escape.h>
 #include <apr_atomic.h>
 #include <apr_portable.h>
 #include <apr_signal.h>
@@ -52,16 +51,14 @@ static const char *CONTENT_TYPE_JSON = "application/json";
 static const char *CONTENT_TYPE_HTML = "text/html";
 
 // constants
-static const char *PERIMETERX_MODULE_VERSION = "Apache Module v2.8.0-rc.9";
+static const char *PERIMETERX_MODULE_VERSION = "Apache Module v2.8.0";
 static const char *SCORE_HEADER_NAME = "X-PX-SCORE";
 static const char *VID_HEADER_NAME = "X-PX-VID";
 static const char *UUID_HEADER_NAME = "X-PX-UUID";
 static const char *ACCEPT_HEADER_NAME = "Accept";
 static const char *ACCESS_CONTROL_ALLOW_ORIGIN_HEADER_NAME = "Access-Control-Allow-Origin";
-static const char *ORIGIN_HEADER_NAME = "Origin";
 static const char *ORIGIN_WILDCARD_VALUE = "*";
 
-static const char *CAPTCHA_COOKIE = "_pxCaptcha";
 static const int MAX_CURL_POOL_SIZE = 10000;
 static const int ERR_BUF_SIZE = 128;
 
@@ -73,6 +70,10 @@ static const char *ERROR_BASE_URL_BEFORE_APP_ID = "mod_perimeterx: BaseUrl was s
 
 static const char *BLOCKED_ACTIVITY_TYPE = "block";
 static const char *PAGE_REQUESTED_ACTIVITY_TYPE = "page_requested";
+
+static const char *LOGGER_DEBUG_FORMAT = "[PerimeterX - DEBUG][%s] - %s";
+static const char *LOGGER_ERROR_FORMAT = "[PerimeterX - ERROR][%s] - %s";
+
 
 #ifdef DEBUG
 extern const char *BLOCK_REASON_STR[];
@@ -111,34 +112,35 @@ static void px_ssl_id_callback(CRYPTO_THREADID *id) {
 
 
 char *create_response(px_config *conf, request_context *ctx) {
-    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, "[%s]: create_response: response creation started", conf->app_id);
+    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "create_response: response creation started");
 
     // support for Access-Control-Allow-Origin headers
     if (conf->origin_wildcard_enabled) {
         apr_table_set(ctx->r->headers_out, ACCESS_CONTROL_ALLOW_ORIGIN_HEADER_NAME,ORIGIN_WILDCARD_VALUE);
-        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, "[%s]: create_response: header Access-Control-Allow-Origin: * set on response", conf->app_id);
+        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "create_response: header Access-Control-Allow-Origin: * set on response");
     } else if (conf->origin_envvar_name) {
         const char *origin_envvar_value = apr_table_get(ctx->r->subprocess_env, conf->origin_envvar_name);
         if (origin_envvar_value != NULL) {
             apr_uri_t origin_envvar_uri;
             apr_status_t origin_envvar_parse_result = apr_uri_parse(ctx->r->pool, origin_envvar_value, &origin_envvar_uri);
             if (origin_envvar_parse_result != 0) {
-                ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, "[%s]: create_response: Origin header was not a valid URI", conf->app_id);
+                ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "create_response: Origin header was not a valid URI");
             } else {
                 // Unparse to ensure there is a real URI
                 const char *unparsed_uri = apr_uri_unparse(ctx->r->pool, &origin_envvar_uri, APR_URI_UNP_OMITPATHINFO);
                 if (unparsed_uri != NULL) {
                     if (strlen(unparsed_uri) > 0) {
-                        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, "[%s]: create_response: unparsed uri  %s" , conf->app_id,unparsed_uri);
+                        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, conf->app_id, apr_pstrcat(ctx->r->pool, "create_response: unparsed uri ", unparsed_uri, NULL));
                         apr_table_set(ctx->r->headers_out, ACCESS_CONTROL_ALLOW_ORIGIN_HEADER_NAME, unparsed_uri);
-                        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, "[%s]: create_response: header Access-Control-Allow-Origin: %s set on response", conf->app_id, origin_envvar_value);
+                        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, conf->app_id, apr_pstrcat(ctx->r->pool, "create_response: header Access-Control-Allow-Origin: ", origin_envvar_value, " set on response", NULL));
+
                     } else {
-                        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, "[%s]: create_response: invalid URI set in envvar" , conf->app_id);
+                        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "create_response: invalid URI set in envvar");
                     }
                 }
             }
         } else {
-            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, "[%s]: create_response: envvar NULL skipped setting Access-Control-Allow-Origin header", conf->app_id);
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "create_response: envvar NULL skipped setting Access-Control-Allow-Origin header");
         }
     }
 
@@ -176,9 +178,9 @@ char *create_response(px_config *conf, request_context *ctx) {
     if (ctx->token_origin == TOKEN_ORIGIN_HEADER) {
         int expected_encoded_len = apr_base64_encode_len(html_size);
         char *encoded_html = apr_palloc(ctx->r->pool, expected_encoded_len + 1);
-        int encoded_len = apr_base64_encode(encoded_html, html, html_size);
+        apr_base64_encode(encoded_html, html, html_size);
         free(html);
-        if (encoded_html == 0) {
+        if (!encoded_html) {
             return NULL;
         }
         return create_mobile_response(conf, ctx, encoded_html);
@@ -191,7 +193,7 @@ void post_verification(request_context *ctx, px_config *conf, bool request_valid
         const char *activity_type = request_valid ? PAGE_REQUESTED_ACTIVITY_TYPE : BLOCKED_ACTIVITY_TYPE;
         char *activity = create_activity(activity_type, conf, ctx);
         if (!activity) {
-            ap_log_error(APLOG_MARK, APLOG_DEBUG, 0, ctx->r->server, "[%s]: post_verification: (%s) create activity failed", ctx->app_id, activity_type);
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, ctx->r->server, LOGGER_DEBUG_FORMAT, ctx->app_id, apr_pstrcat(ctx->r->pool, "post_verification: ", activity_type, " create activity failed", NULL));
             return;
         }
         if (conf->background_activity_send) {
@@ -211,7 +213,7 @@ int px_handle_request(request_rec *r, px_config *conf) {
 
     // Decline internal redirects and subrequests
     if (r->prev) {
-            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, "[%s]: px_handle_request: request declined - interal redirect or subrequest", conf->app_id);
+        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "Request declined - interal redirect or subrequest");
 	    return DECLINED;
     }
 
@@ -222,18 +224,21 @@ int px_handle_request(request_rec *r, px_config *conf) {
     if (conf->skip_mod_by_envvar) {
         const char *skip_px = apr_table_get(r->subprocess_env, "PX_SKIP_MODULE");
         if  (skip_px != NULL) {
-            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, "[%s]: px_handle_request: PX_SKIP_MODULE was set on the request", conf->app_id);
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "Request will not be verified, module is disabled by EnvIf");
             return DECLINED;
         }
     }
 
+    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "Starting request verification");
+
     request_context *ctx = create_context(r, conf);
     if (ctx) {
+        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "Request context created successfully");
         bool request_valid = px_verify_request(ctx, conf);
 
         // if request is not valid, and monitor mode is on, toggle request_valid and set pass_reason
         if (conf->monitor_mode && !request_valid) {
-            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, "[%s]: request marked for simulated block", conf->app_id);
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "Request marked for simulated block");
             ctx->pass_reason = PASS_REASON_MONITOR_MODE;
             request_valid = true;
         }
@@ -254,23 +259,12 @@ int px_handle_request(request_rec *r, px_config *conf) {
             apr_table_set(r->headers_in, conf->score_header_name, score_str);
         }
 
-        ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, "[%s]: request_valid %d , block_enabled %d ", conf->app_id, request_valid, ctx->block_enabled);
-
         if (!request_valid && ctx->block_enabled) {
             // redirecting requests to custom block page if exists
             if (conf->block_page_url) {
-                const char *redirect_url;
-                const char *url_arg = r->args
-                    ? apr_pstrcat(r->pool, r->uri, "?", r->args, NULL)
-                    : apr_pstrcat(r->pool, r->uri, NULL);
-                apr_size_t encoded_url_len = 0;
-                if (apr_escape_urlencoded(NULL, url_arg, APR_ESCAPE_STRING, &encoded_url_len) == APR_SUCCESS)   {
-                    char *encoded_url = apr_pcalloc(r->pool,encoded_url_len + 1);
-                    apr_escape_urlencoded(encoded_url, url_arg, APR_ESCAPE_STRING, NULL);
-                    redirect_url = apr_pstrcat(r->pool, conf->block_page_url, "?url=", encoded_url, "&uuid=", ctx->uuid, "&vid=", ctx->vid,  NULL);
-                } else {
-                    redirect_url = apr_pstrcat(r->pool, conf->block_page_url, "?url=", r->uri, "&uuid=", ctx->uuid, "&vid=", ctx->vid,  NULL);
-                }
+                const char *url_arg = r->args ? apr_pstrcat(r->pool, r->uri, "?", r->args, NULL) : r->uri;
+                const char *url = pescape_urlencoded(r->pool, url_arg);
+                const char *redirect_url = apr_pstrcat(r->pool, conf->block_page_url, "?url=", url, "&uuid=", ctx->uuid, "&vid=", ctx->vid,  NULL);
                 apr_table_set(r->headers_out, "Location", redirect_url);
                 return HTTP_TEMPORARY_REDIRECT;
             }
@@ -288,11 +282,10 @@ int px_handle_request(request_rec *r, px_config *conf) {
                 return DONE;
             }
             // failed to create response
-            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, "[%s]: Could not create block page with template, passing request", conf->app_id);
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, LOGGER_DEBUG_FORMAT, conf->app_id, "Could not create block page with template, passing request");
         }
     }
     r->status = HTTP_OK;
-    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, r->server, "[%s]: px_handle_request: request passed, score %d, monitor mode %d", ctx->app_id, ctx->score, conf->monitor_mode);
     return OK;
 }
 
@@ -315,12 +308,12 @@ static void *APR_THREAD_FUNC health_check(apr_thread_t *thd, void *data) {
 
         apr_thread_mutex_unlock(conf->health_check_cond_mutex);
         if (conf->should_exit_thread) {
-            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, hc->server, "health_check: marked to exit");
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, hc->server, LOGGER_DEBUG_FORMAT, conf->app_id, "health_check: marked to exit");
             break;
         }
 
         // do health check until success
-        CURLcode res = CURLE_AGAIN;
+        res = CURLE_AGAIN;
         while (!conf->should_exit_thread && res != CURLE_OK) {
             curl_easy_setopt(curl, CURLOPT_URL, health_check_url);
             curl_easy_setopt(curl, CURLOPT_TIMEOUT_MS, conf->api_timeout_ms);
@@ -332,7 +325,8 @@ static void *APR_THREAD_FUNC health_check(apr_thread_t *thd, void *data) {
         apr_atomic_set32(&conf->px_errors_count, 0);
     }
 
-    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, hc->server, "health_check: thread exiting");
+    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, hc->server, LOGGER_DEBUG_FORMAT, conf->app_id, "health_check: thread exiting");
+
     curl_easy_cleanup(curl);
     apr_thread_exit(thd, 0);
     return NULL;
@@ -345,7 +339,7 @@ static void *APR_THREAD_FUNC background_activity_consumer(apr_thread_t *thd, voi
 
     void *v;
     if (!curl) {
-        ap_log_error(APLOG_MARK, LOG_ERR, 0, consumer_data->server, "[%s]: could not create curl handle, thread will not run to consume messages", conf->app_id);
+        ap_log_error(APLOG_MARK, APLOG_ERR, 0, consumer_data->server, LOGGER_DEBUG_FORMAT, conf->app_id, "could not create curl handle, thread will not run to consume messages");
         return NULL;
     }
 
@@ -365,10 +359,9 @@ static void *APR_THREAD_FUNC background_activity_consumer(apr_thread_t *thd, voi
     }
 
     curl_easy_cleanup(curl);
-    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, consumer_data->server,
-            "[%s]: activity consumer thread exited", conf->app_id);
+    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, consumer_data->server, LOGGER_DEBUG_FORMAT, conf->app_id, "activity consumer thread exited");
     apr_thread_exit(thd, 0);
-    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, consumer_data->server, "[%s]: Sending activity completed", conf->app_id);
+    ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, consumer_data->server, LOGGER_DEBUG_FORMAT, conf->app_id, "Sending activity completed");
     return NULL;
 }
 
@@ -385,19 +378,19 @@ static apr_status_t create_health_check(apr_pool_t *p, server_rec *s, px_config 
 
     rv = apr_thread_cond_create(&cfg->health_check_cond, p);
     if (rv != APR_SUCCESS) {
-        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, "error while init health_check thread cond");
+        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "error while init health_check thread cond");
         return rv;
     }
 
     rv = apr_thread_create(&cfg->health_check_thread, NULL, health_check, (void*) hc_data, p);
     if (rv != APR_SUCCESS) {
-        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, "error while init health_check thread create");
+        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "error while init health_check thread create");
         return rv;
     }
 
     rv = apr_thread_mutex_create(&cfg->health_check_cond_mutex, 0, p);
     if (rv != APR_SUCCESS) {
-        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, "error while creating health_check thread mutex");
+        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "error while creating health_check thread mutex");
         return rv;
     }
 
@@ -409,8 +402,7 @@ static apr_status_t background_activity_send_init(apr_pool_t *pool, server_rec *
 
     rv = apr_queue_create(&cfg->activity_queue, cfg->background_activity_queue_size, pool);
     if (rv != APR_SUCCESS) {
-        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s,
-                "[%s]: failed to initialize background activity queue", cfg->app_id);
+        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "failed to initialize background activity queue");
         return rv;
     }
 
@@ -420,20 +412,19 @@ static apr_status_t background_activity_send_init(apr_pool_t *pool, server_rec *
 
     rv = apr_thread_pool_create(&cfg->activity_thread_pool, 0, cfg->background_activity_workers, pool);
     if (rv != APR_SUCCESS) {
-        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s,
-                "[%s]: failed to initialize background activity thread pool", cfg->app_id);
+        ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "failed to initialize background activity thread pool");
         return rv;
     }
 
     for (unsigned int i = 0; i < cfg->background_activity_workers; ++i) {
         rv = apr_thread_pool_push(cfg->activity_thread_pool, background_activity_consumer, consumer_data, 0, NULL);
         if (rv != APR_SUCCESS) {
-            ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, "failed to push background activity consumer");
+            ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "failed to push background activity consumer");
             return rv;
         }
     }
 
-    ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, "finished init background activities");
+    ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "finished init background activities");
     return rv;
 }
 
@@ -448,8 +439,9 @@ static apr_status_t px_child_exit(void *data) {
         apr_thread_cond_signal(cfg->health_check_cond);
     }
     // terminate the queue and wake up all idle threads
+    apr_status_t rv = APR_SUCCESS;
     if (cfg->activity_queue) {
-        apr_status_t rv = apr_queue_term(cfg->activity_queue);
+        rv = apr_queue_term(cfg->activity_queue);
         if (rv != APR_SUCCESS) {
             char buf[ERR_BUF_SIZE];
             char *err = apr_strerror(rv, buf, sizeof(buf));
@@ -457,6 +449,7 @@ static apr_status_t px_child_exit(void *data) {
         }
     }
     ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, s, "px_child_exit: cleanup finished");
+    return rv;
 }
 
 static apr_status_t px_child_setup(apr_pool_t *p, server_rec *s) {
@@ -466,36 +459,35 @@ static apr_status_t px_child_setup(apr_pool_t *p, server_rec *s) {
 
         px_config *cfg = ap_get_module_config(vs->module_config, &perimeterx_module);
         if (!cfg || !cfg->module_enabled) {
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, s, LOGGER_DEBUG_FORMAT, cfg->app_id, "Request will not be verified, module is disabled");
             continue;
         }
         // initialize the PerimeterX needed pools and background workers if the PerimeterX module is enabled
 
         rv = apr_pool_create(&cfg->pool, vs->process->pool);
         if (rv != APR_SUCCESS) {
-            ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, "px_child_setup: error while trying to initialize apr_pool for configuration");
+            ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "px_child_setup: error while trying to initialize apr_pool for configuration");
             return rv;
         }
 
         cfg->curl_pool = curl_pool_create(cfg->pool, cfg->curl_pool_size);
 
         if (cfg->background_activity_send) {
-            ap_log_error(APLOG_MARK, APLOG_CRIT | APLOG_NOERRNO, 0, s,
-                    "px_child_setup: start init for background_activity_send");
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, s, LOGGER_DEBUG_FORMAT, cfg->app_id, "px_child_setup: start init for background_activity_send");
+
             rv = background_activity_send_init(cfg->pool, vs, cfg);
             if (rv != APR_SUCCESS) {
-                ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s,
-                        "px_child_setup: error while trying to init background_activity_consumer");
+                ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "px_child_setup: error while trying to init background_activity_consumer");
                 return rv;
             }
         }
 
         if (cfg->px_health_check) {
-            ap_log_error(APLOG_MARK, APLOG_CRIT | APLOG_NOERRNO, 0, s,
-                    "px_child_setup: setting up health_check thread");
+            ap_log_error(APLOG_MARK, APLOG_DEBUG | APLOG_NOERRNO, 0, s, LOGGER_DEBUG_FORMAT, cfg->app_id, "px_child_setup: setting up health_check thread");
+
             rv = create_health_check(cfg->pool, vs, cfg);
             if (rv != APR_SUCCESS) {
-                ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s,
-                        "px_child_setup: error while trying to init health_check_thread");
+                ap_log_error(APLOG_MARK, APLOG_CRIT, rv, s, LOGGER_ERROR_FORMAT, cfg->app_id, "px_child_setup: error while trying to init health_check_thread");
                 return rv;
             }
         }
@@ -744,7 +736,6 @@ static const char *set_block_page_url(cmd_parms *cmd, void *config, const char *
 }
 
 static const char *add_route_to_whitelist(cmd_parms *cmd, void *config, const char *route) {
-    const char *sep = ";";
     px_config *conf = get_config(cmd, config);
     if (!conf) {
         return ERROR_CONFIG_MISSING;