+++ src/px_types.h
@@ -94,6 +94,7 @@ typedef enum {
     VALIDATION_RESULT_DECRYPTION_FAILED,
     VALIDATION_RESULT_NULL_PAYLOAD,
     VALIDATION_RESULT_MOBILE_SDK_CONNECTION_ERROR,
+    VALIDATION_RESULT_MOBILE_SDK_PINNING_ERROR
 } validation_result_t;
 
 typedef enum call_reason_t {
@@ -105,6 +106,7 @@ typedef enum call_reason_t {
     CALL_REASON_SENSITIVE_ROUTE,
     CALL_REASON_CAPTCHA_FAILED,
     CALL_REASON_MOBILE_SDK_CONNECTION_ERROR,
+    CALL_REASON_MOBILE_SDK_PINNING_ERROR
 } call_reason_t;
 
 typedef enum {
@@ -133,6 +135,7 @@ typedef enum {
 typedef enum {
     ACTION_CAPTCHA,
     ACTION_BLOCK,
+    ACTION_CHALLENGE,
 } action_t;
 
 typedef struct risk_payload_t {
@@ -154,6 +157,7 @@ typedef struct risk_response_t {
     int status;
     int score;
     const char *action;
+    const char *action_data_body;
 } risk_response;
 
 typedef struct captcha_response_t {
@@ -194,6 +198,7 @@ typedef struct request_context_t {
     token_origin_t token_origin;
     action_t action;
     bool response_application_json;
+    const char *action_data_body;
 } request_context;
 
 typedef enum {