+++ CHANGELOG.md
@@ -4,9 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.8.0] - 01-22-2018
+* Added support for APR 1.4.6
+* Added support for js challenge
+* Added new logs format
+ -- RC Changes Below
+* Added CORS support
+* Added Monitor Mode, default mode is Off
+* Added support new captcha flow
+* Added support Cookie V3
+* Added support for funCaptcha
+* Modified rendering block pages by action
+* Fixed internal request handling
+* Fixed mobile sdk and configurations
+* Improved stability and memory allocations 
+* Improved SSL Optimizations
+
 ## [2.8.0-rc.9] 11-20-2017
 * Init and clean OpenSSL with threading support 
-* Switched background activitieis to off by default\
+* Switched background activities to off by default\
 * Added debug symbols to debian installation
 * Fixed reCaptcha support for mobile sdk
 * Memory leak optimizations